Set up the argparse and logging modules
The argparse module has been set up to parse arguments from the command
line and provide defaults for some needed variables.

The logging implementation was changed to use both a file and a stream handler.

As part of this process, many of the previously defined functions were
removed. That's okay, not needed really and this implementation will be
much better in the long run.